+++ src/Ext/AnimType/Body.h
@@ -28,6 +28,9 @@ public:
 		Valueable<int> HideIfNoOre_Threshold;
 		Nullable<bool> Layer_UseObjectLayer;
 		Valueable<bool> UseCenterCoordsIfAttached;
+		Nullable<WeaponTypeClass*> Weapon;
+		Valueable<int> Damage_Delay;
+		Valueable<bool> Damage_DealtByOwner;
 
 		ExtData(AnimTypeClass* OwnerObject) : Extension<AnimTypeClass>(OwnerObject)
 			, Palette { CustomPalette::PaletteMode::Temperate }
@@ -43,6 +46,9 @@ public:
 			, HideIfNoOre_Threshold { 0 }
 			, Layer_UseObjectLayer {}
 			, UseCenterCoordsIfAttached { false }
+			, Weapon {}
+			, Damage_Delay { 0 }
+			, Damage_DealtByOwner { false }
 		{ }
 
 		virtual ~ExtData() = default;